Working drag and drop
Working drag and drop implementation (at least as far as
we had it working previously) can be found in app.html,
drag-drop-app.js, and drag-drop.css.

As before, it is just dragging and dropping between lists
and there is still the (CSS?) issue where an empty list
can't have anything dropped onto it.

We'll need to port these files into the appropriate folders
and remove that IIFE style in the drag-drop-app.js file that
is not a good Angular practice.

Best of luck.