localStorage prototype done, moving to JWT implementation
Currently users can be created and updated using proper
authentication headers. The POST Sessions route can also
be hit, but it doesn't actually create a session; it just
returns a JSON of user, which is misleading and will not
allow a user to login if they accidentally wipe their
local storage of their user information (since they need
it for the auth headers).

Implementing JWT will allow users hitting the Sessions
route to generate a JWT if they don't send one in the
POST '/sessions' request. That way if their storage
ever gets cleared, the login will act like an actual
login, otherwise they shouldn't need to login because
the isAuthed function should prevent them from being
redirected to the auth path.

In short, the flow should be:

Register as a user. Create a JWT in the API and return
the encoded token. Client stores token in localStorage.

All future API calls pass the token in their headers.
API decodes and compares the received token against the
user's token stored in the database.
If the API ever receives a call requiring authentication
that has an invalid or missing token, it gives a 401.

Version A:
Client does not log out. Token stays in localStorage.
Client is able to make API calls.

Version B:
Client logs out, deleting token from localStorage.
On next visit, client logs in. API generates a new
JWT for them (if they are a valid user) and returns
it and their information. Client can make API calls.